Remove dependency requirement for ACF Plugin and add options to editor
I removed the the requirement for needing the Advanced Custom Fields plugin by moving those fields to metaboxes. Also added support for configuring the header through a new options page under the Appearance section of the back-end. You can now change the Site Title that is displayed n the header, with an separate field for the top and bottom lines as well as the logo displayed there. The Landing Page template was also made the default page design and options given to control the hero header including selecting an image for it. Options for the left navigation menu were also moved to a metabox on the edit page.